+++ joyrpc-plugin/joyrpc-protocol/joyrpc-protocol-grpc/src/main/java/io/joyrpc/protocol/grpc/GrpcServerProtocol.java
@@ -82,7 +82,7 @@ public class GrpcServerProtocol extends AbstractProtocol implements ServerProtoc
     @Override
     protected void onException(final Channel channel, final MessageHeader header, final RpcException cause) {
         int streamId = (Integer) header.getAttributes().get(HeaderMapping.STREAM_ID.getNum());
-        int msgId = header.getMsgId();
+        int msgId = (int) header.getMsgId();
         Http2Headers endHeaders = Headers.build(cause);
         Http2ResponseMessage message = new DefaultHttp2ResponseMessage(streamId, msgId, null, null, endHeaders);
         channel.send(message, (event -> {
